+++ index.html
@@ -1,6 +1,7 @@
 <!DOCTYPE html>
 <html lang="ru">
 <head>
+  <script src="main.js" type="module"></script>
   <script src="js/functions.js"></script>
   <script src="js/main.js"></script>
   <meta charset="utf-8">
